Rename Lite-On commands to ReadBuffer3C
