+++ Aaru.Core/Devices/Dumping/Sbc/Trim.cs
@@ -99,7 +99,7 @@ partial class Dump
             _resume.BadBlocks.Remove(badSector);
             extents.Add(badSector);
 
-            if(scsiReader.LiteOnReadRaw || scsiReader.HldtstReadRaw)
+            if(scsiReader.ReadBuffer3CReadRaw || scsiReader.HldtstReadRaw)
             {
                 var cmi = new byte[1];
